refractored header component
