+++ src/Components/Header/Nav.js
@@ -1,0 +1,21 @@
+import React from 'react';
+
+import {
+  MobileMenu,
+  NavTriangle,
+  StyledLink,
+} from '../../Styles/Header/NavStyles.js';
+
+const Nav = () => {
+  return (
+    <MobileMenu>
+      <NavTriangle/>
+      <StyledLink to="/">Home</StyledLink>
+      <StyledLink to="/">About</StyledLink>
+      <StyledLink to="/">Community</StyledLink>
+      <StyledLink to="/">Contact</StyledLink>
+    </MobileMenu>
+  );
+};
+
+export default Nav;
